Move pagination logics to client
